now one logger
