+++ seldon-request-logger/Makefile
@@ -4,28 +4,15 @@ build_image:
 push_image:
 	docker push seldonio/seldon-request-logger:latest
 
-build_image_cifar:
-	docker build --no-cache -f cifar10.Dockerfile . -t seldonio/seldon-request-logger-cifar:latest
-
-push_image_cifar:
-	docker push seldonio/seldon-request-logger-cifar:latest
-
 run_container:
 	docker run -p 2222:8080 seldonio/seldon-request-logger:latest
-
-run_container_cifar:
-	docker run -p 2222:8080 seldonio/seldon-request-logger-cifar:latest
 
 run_local: export FLASK_RUN_PORT = 2222
 run_local: export FLASK_APP = default_logger.py
 run_local:
 	cd app && flask run
 
-run_local_cifar: export FLASK_RUN_PORT = 2222
-run_local_cifar: export FLASK_APP = cifar_logger.py
-run_local_cifar:
-	cd app && flask run
-
+#see test.sh for more test data
 test:
 	curl -v "http://localhost:2222/" \
       -X POST \
