+++ seldon-request-logger/README.md
@@ -4,16 +4,43 @@ For use with request logging example (/examples/centralised-logging/request-logg
 
 This example request logger is as general-purpose as possible. Eventually intended to read all types of CloudEvents emitted by the Seldon executor. These can be:
 
- - GRPC or json
  - SeldonMessages or tensorflow format
  - Tabular data, string data or image data
- - Requests, responses to main predictor or to canary or shadow (all to be aggregated)
+ - Requests, responses to main predictor or to canary or shadow (all to be recorded in traceable way)
+ - GRPC or json
 
-Custom request loggers can be built for different types of transformations. Can be written in any language, just needs to handle HTTP POST requests and log to stdout for fluentd or could go direct to chosen backend.
+Custom request loggers can be built for different types of transformations. Can be written in any language, just needs to handle HTTP POST requests containing cloud events and can log to chosen backend.
 
-A custom logger example is provided for a [CIFAR10 image classfier example](https://github.com/SeldonIO/alibi-detect/tree/master/integrations/samples/kfserving/od-cifar10)
+# Design
+
+A SeldonDeployment or KFServing InferenceService is designed to output a dump of a request or response via http. Each includes headers indicating the source and the request id.
+
+Different models and different endpoints are logged to different indexes. This ensures that models with different data formats won't conflict in document formats.
+
+Different endpoints in different indexes also allows us to handle shadows. A shadow of a model will receive the same request but may have a different output. So we log both sets under different indexes.
+
+Batch requests are split and indexed as separate documents with `-item-N` appended to the doc id.
+
+See comments in log_helper.py for details of headers.
 
 # Local Testing
+First run elastic locally
+```
+docker pull docker.elastic.co/elasticsearch/elasticsearch-oss:7.6.0
+docker run -p 9200:9200 -p 9300:9300 -e "discovery.type=single-node" docker.elastic.co/elasticsearch/elasticsearch-oss:7.6.0
+```
+Run the logger:
+```
+make run_local
+```
+And in another window run app/test.sh
+
+The output of the logger will say where docs are indexed. The contents can be checked in postman by querying on the elastic host e.g.
+```
+localhost:9200/inference-log-seldon-unknown-namespace-strdata-unknown-endpoint/inferencerequest/7g
+```
+
+# Local Testing Against Seldon Executor
 
 To try this out, run elastic locally
 ```
@@ -92,10 +119,8 @@ Example output is:
 
 # On-going work
 
-TODO: DOCUMENT HEADERS AND EXTENSION ATTRIBUTES BETTER
 TODO: SOURCE IS ALWAYS http://localhost:8000/ WHEN COMING FROM EXECUTOR
 
-TODO: BATCH IS BROKEN BY THIS WAY OF USING REQ IDS - NOW THE SECOND ROW OVERWRITES THE FIRST. MAY HAVE TO ADD ORDINAL TO DOC ID AND ENSURE ORIGINAL ID IN DOC BODY.
-TODO: HANDLE GRPC AND INFERENCESERVICES
+TODO: HANDLE GRPC
 TODO: THINK ABOUT SHADOW CASE - HOW TO ENSURE WE HAVE SOMETHING TO LINK DEFAULT AND SHADOW? MULTIPLE SHADOWS? https://github.com/SeldonIO/seldon-core/issues/1207
 TODO: FEEDBACK - IF SENT WITH CUSTOM ID HEADER COULD SUPPORT A/B TESTS WITH RECORDED RESULTS
